+++ src/editor/controllers/selection-controller.tsx
@@ -2,8 +2,8 @@ import { createEffect, createRoot, createSignal } from "solid-js";
 import { EditorSpace } from "~/editor/internal/editor-space";
 import { Vector2 } from "~/wrapper/data_types/geometry";
 import { GraphNode } from "~/wrapper/nodes/graph-node";
-import { Grid } from "../ui/misc/grid";
-import { SelectionRect } from "../ui/misc/selection_rect";
+import { Grid } from "../ui/editor/components/misc/grid";
+import { SelectionRect } from "../ui/editor/components/misc/selection_rect";
 import { NodeSlot } from "~/wrapper/nodes/slot/node-slot";
 import { ConnectionController } from "~/wrapper/controllers/connection-controller";
 import { ConnActionUtils } from '../../network/controllers/actions/conn-actions';
