+++ src/editor/ui/editor/components/misc/selection_rect.tsx
