+++ src/editor/ui/editor/components/node/node-field.tsx
@@ -1,10 +1,10 @@
-import { GraphNode } from "../../../wrapper/nodes/graph-node";
-import { NodeParameter } from "../../../wrapper/nodes/data/node-data";
-import { DefaultDataTypes } from "../../../wrapper/nodes/data/node-data-type";
+import { GraphNode } from "../../../../../wrapper/nodes/graph-node";
+import { NodeParameter } from "../../../../../wrapper/nodes/data/node-data";
+import { DefaultDataTypes } from "../../../../../wrapper/nodes/data/node-data-type";
 import { createMemo, For, Match, Show, Switch } from "solid-js";
 import { debounce } from "~/editor/utils/debounce-utils";
 import { UserWorkspace } from "~/network/session/user-workspace";
-import { clamp } from '../../utils/parameter-utils';
+import { clamp } from '../../../../utils/parameter-utils';
 import { session_controller } from "~/singletons/user_session";
 
 const TYPE_CONFIGS: Record<string, { type: string; step?: number; min?: number }> = {
