+++ src/editor/ui/editor/components/node/output/array-output.tsx
