+++ src/editor/ui/editor/components/node/output/node-output.tsx
