+++ src/editor/ui/editor/components/node/output/scalar-output.tsx
