+++ src/editor/ui/editor/components/node/slot-components.tsx
