+++ src/editor/ui/editor/components/node/slot-output.tsx
