+++ src/editor/ui/screens/settings-view.tsx
@@ -1,0 +1,3 @@
+export const SettingsView = (props: {}) => {
+    
+}
